+++ src/aui-scheduler/js/aui-scheduler-base.js
@@ -11,6 +11,7 @@ var CSS_SCHEDULER_VIEW_ = A.getClassName('scheduler-base', 'view', ''),
     DateMath = A.DataType.DateMath,
     Lang = A.Lang,
     isArray = Lang.isArray,
+    isBoolean = Lang.isBoolean,
     isDate = Lang.isDate,
     isFunction = Lang.isFunction,
     isNumber = Lang.isNumber,
@@ -774,6 +775,18 @@ var SchedulerBase = A.Component.create({
         },
 
         /**
+         * Define whether the scheduler header will be displayed.
+         *
+         * @attribute showHeader
+         * @default true
+         * @type {Boolean}
+         */
+        showHeader: {
+            validator: isBoolean,
+            value: true
+        },
+
+        /**
          * Contains the node for the select dropdown for `Scheduler`'s views.
          * This node is only visible on mobile.
          *
@@ -858,7 +871,7 @@ var SchedulerBase = A.Component.create({
      * @type {Array}
      * @static
      */
-    UI_ATTRS: ['date', 'activeView'],
+    UI_ATTRS: ['date', 'activeView', 'showHeader'],
 
     /**
      * Static property used to define the augmented classes.
@@ -894,6 +907,8 @@ var SchedulerBase = A.Component.create({
             instance.viewsSelectNode = instance.get('viewsSelectNode');
             instance.todayNode = instance.get('todayNode');
             instance.viewsNode = instance.get('viewsNode');
+
+            instance._populateViewNodes();
 
             instance.after({
                 activeViewChange: instance._afterActiveViewChange,
@@ -1064,12 +1079,14 @@ var SchedulerBase = A.Component.create({
         renderButtonGroup: function() {
             var instance = this;
 
-            instance.buttonGroup = new A.ButtonGroup({
-                boundingBox: instance.viewsNode,
-                on: {
-                    selectionChange: A.bind(instance._onButtonGroupSelectionChange, instance)
-                }
-            }).render();
+            if (!instance.buttonGroup) {
+                instance.buttonGroup = new A.ButtonGroup({
+                    boundingBox: instance.viewsNode,
+                    on: {
+                        selectionChange: A.bind(instance._onButtonGroupSelectionChange, instance)
+                    }
+                }).render();
+            }
         },
 
         /**
@@ -1090,26 +1107,29 @@ var SchedulerBase = A.Component.create({
          */
         syncStdContent: function() {
             var instance = this;
-            var views = instance.get('views');
 
-            instance.navNode.append(instance.iconPrevNode);
-            instance.navNode.append(instance.todayNode);
-            instance.navNode.append(instance.iconNextNode);
+            if (instance.get('showHeader')) {
+                instance.renderButtonGroup();
 
-            instance.controlsNode.append(instance.navNode);
-            instance.controlsNode.append(instance.navDateNode);
+                instance.navNode.append(instance.iconPrevNode);
+                instance.navNode.append(instance.todayNode);
+                instance.navNode.append(instance.iconNextNode);
 
-            A.Array.each(views, function(view) {
-                instance.viewsSelectNode.append(instance._createViewTriggerNode(view, TPL_SCHEDULER_VIEW_LIST));
-                instance.viewsNode.append(instance._createViewTriggerNode(view, TPL_SCHEDULER_VIEW_BUTTON));
-            });
+                instance.controlsNode.append(instance.navNode);
+                instance.controlsNode.append(instance.navDateNode);
 
-            instance.viewsNode.append(instance.viewsSelectNode);
+                instance.viewsNode.append(instance.viewsSelectNode);
 
-            instance.header.append(instance.controlsNode);
-            instance.header.append(instance.viewsNode);
+                instance.header.append(instance.controlsNode);
+                instance.header.append(instance.viewsNode);
 
-            instance.setStdModContent(WidgetStdMod.HEADER, instance.header.getDOM());
+                instance.setStdModContent(WidgetStdMod.HEADER, instance.header.getDOM());
+
+                instance.header.show();
+            }
+            else {
+                instance.header.hide();
+            }
         },
 
         /**
@@ -1154,7 +1174,6 @@ var SchedulerBase = A.Component.create({
                 activeView = instance.get('activeView');
 
             instance.renderView(activeView);
-            instance.renderButtonGroup();
             instance.renderDropdownList();
 
             instance._uiSetDate(instance.get('date'));
@@ -1318,6 +1337,32 @@ var SchedulerBase = A.Component.create({
         },
 
         /**
+         * Add the content to represent each view to the corresponding view
+         * nodes.
+         *
+         * The scheduler can display its events in many views - day view, week
+         * view etc. These views are selected through a set of buttons at the
+         * top left of the scheduler (if it is being viewed in desktop mode)
+         * or from a select (if it is being viewed from a mobile device). This
+         * method checks which views the scheduler is expected to render, and
+         * adds elements both options (i.e. adds buttons to the button group and
+         * options to the select).
+         *
+         * @method _populateViewNodes
+         * @protected
+         */
+        _populateViewNodes: function() {
+            var instance = this;
+
+            var views = instance.get('views');
+
+            A.Array.each(views, function(view) {
+                instance.viewsSelectNode.append(instance._createViewTriggerNode(view, TPL_SCHEDULER_VIEW_LIST));
+                instance.viewsNode.append(instance._createViewTriggerNode(view, TPL_SCHEDULER_VIEW_BUTTON));
+            });
+        },
+
+        /**
          * Plugs 'NodeFocusManager' into the viewsNode and navNode.
          *
          * @method _plugFocusManager
@@ -1447,6 +1492,19 @@ var SchedulerBase = A.Component.create({
 
                 instance.syncEventsUI();
             }
+        },
+
+        /**
+         * Set the `showHeader` property - i.e., defines if the scheduler
+         * header will be displayed.
+         *
+         * @method _uiSetShowHeader
+         * @protected
+         */
+        _uiSetShowHeader: function() {
+            var instance = this;
+
+            instance.syncStdContent();
         }
     }
 });