+++ src/aui-tooltip/tests/unit/js/aui-tooltip-delegate-tests.js
@@ -82,7 +82,7 @@ YUI.add('aui-tooltip-delegate-tests', function(Y) {
 
                 this.wait(function() {
                     Y.Assert.isFalse(tooltip.getStyle('opacity') > 0, '.tooltip is not hidden.');
-        
+
                     trigger.simulate('focus');
 
                     this.wait(function() {
@@ -108,10 +108,64 @@ YUI.add('aui-tooltip-delegate-tests', function(Y) {
 
             Y.Assert.areEqual('mouseenter', this._delegate.get('triggerShowEvent'));
             Y.Assert.areEqual('mouseleave', this._delegate.get('triggerHideEvent'));
+        },
+
+        'tooltips should be aligned to trigger': function() {
+            var tooltip1 = Y.one('#trigger1');
+            var tooltip2 = Y.one('#trigger2');
+
+            var tooltip = new Y.TooltipDelegate(
+                {
+                    position: 'right',
+                    trigger: '.tooltip-delegate li',
+                    triggerHideEvent: ['blur', 'mouseleave'],
+                    triggerShowEvent: ['focus', 'mouseover'],
+                    visible: false
+                }
+            );
+
+            tooltip1.simulate('mouseover');
+
+            tooltip = tooltip.getTooltip();
+
+            var contentBox = tooltip.get('contentBox');
+
+            var oldTooltipContent = contentBox.text();
+            var oldOffsetLeft = contentBox.get('offsetLeft');
+            var oldOffsetTop = contentBox.get('offsetTop');
+
+            tooltip1.simulate('mouseout');
+
+            tooltip2.simulate('focus');
+            tooltip2.simulate('blur');
+
+            tooltip1.simulate('focus');
+
+            var newTooltipContent = contentBox.text();
+            var newOffsetLeft = contentBox.get('offsetLeft');
+            var newOffsetTop = contentBox.get('offsetTop');
+
+            Y.Assert.areEqual(
+                oldTooltipContent,
+                newTooltipContent,
+                'Tooltip should use the correct content.'
+            );
+
+            Y.Assert.areEqual(
+                oldOffsetTop,
+                newOffsetTop,
+                'Tooltip should be aligned with trigger\'s top position.'
+            );
+
+            Y.Assert.areEqual(
+                oldOffsetLeft,
+                newOffsetLeft,
+                'Tooltip should be aligned with trigger\'s left position.'
+            );
         }
     }));
     Y.Test.Runner.add(suite);
 
 }, '', {
-    requires: ['aui-tooltip-delegate', 'node-event-simulate', 'test', 'node-base']
+    requires: ['aui-tooltip-delegate', 'node-base', 'node-event-simulate', 'test']
 });