+++ src/aui-video/js/aui-video.js
@@ -14,7 +14,7 @@ var Lang = A.Lang,
 
     DOC = A.config.doc,
 
-    TPL_VIDEO = '<video id="{id}" controls="controls" class="' + CSS_VIDEO_NODE + '" {height} {width}></video>',
+    TPL_VIDEO = '<video id="{id}" controls="controls" class="' + CSS_VIDEO_NODE + '"></video>',
     TPL_VIDEO_FALLBACK = '<div class="' + CSS_VIDEO_NODE + '"></div>';
 
 /**
@@ -192,6 +192,19 @@ var Video = A.Component.create({
     prototype: {
 
         /**
+          * Destructor implementation.
+          * Lifecycle.
+          *
+          * @method destructor
+          * @protected
+          */
+        destructor: function() {
+            var instance = this;
+
+            (new A.EventHandle(instance._eventHandles)).detach();
+        },
+
+        /**
          * Render the Video component instance. Lifecycle.
          *
          * @method renderUI
@@ -215,6 +228,8 @@ var Video = A.Component.create({
                     cropType: event.type
                 });
             });
+
+            instance._setResponsiveDimensions();
         },
 
         /**
@@ -234,6 +249,13 @@ var Video = A.Component.create({
 
             instance.publish('play');
             instance.publish('pause');
+
+            instance._eventHandles = [
+                A.after(
+                    'windowresize',
+                    A.bind('_afterWindowResize', instance)
+                )
+            ];
         },
 
         /**
@@ -290,6 +312,19 @@ var Video = A.Component.create({
             if (instance._video.hasMethod('play')) {
                 instance._video.invoke('play');
             }
+        },
+
+        /**
+         * Fired after the `windowresize` event.
+         *
+         * @method _afterWindowResize
+         * @protected
+         */
+        _afterWindowResize: function() {
+            var instance = this;
+
+            instance._responsiveBoundingBox();
+            instance._setResponsiveDimensions();
         },
 
         /**
@@ -388,50 +423,107 @@ var Video = A.Component.create({
          * @protected
          */
         _renderVideo: function(fallback) {
-            var instance = this,
-                attrHeight,
-                attrWidth,
+            var instance,
                 height,
                 tpl,
                 tplObj,
                 video,
                 width;
 
+            instance = this;
             tpl = TPL_VIDEO;
+
+            height = instance.get('height');
+            width = instance.get('width');
 
             if (UA.gecko && fallback) {
                 tpl = TPL_VIDEO_FALLBACK;
             }
-            else {
-                attrHeight = '';
-                attrWidth = '';
-
-                height = instance.get('height');
-
-                width = instance.get('width');
-
-                if (height) {
-                    attrHeight = 'height="' + height + '"';
-                }
-
-                if (width) {
-                    attrWidth = 'width="' + width + '"';
-                }
-            }
 
             tplObj = Lang.sub(
                 tpl, {
-                    height: attrHeight,
-                    id: A.guid(),
-                    width: attrWidth
+                    id: A.guid()
                 }
             );
 
             video = A.Node.create(tplObj);
 
+            if (width) {
+                video.width(width);
+            }
+            if (height) {
+                video.height(height);
+            }
+
             instance.get('contentBox').append(video);
 
             instance._video = video;
+        },
+
+        /**
+         * Remove the defined height and width from the bounding box.
+         *
+         * @method _responsiveBoundingBox
+         * @protected
+         */
+        _responsiveBoundingBox: function() {
+            var instance = this,
+                boundingBox = instance.get('boundingBox');
+
+            boundingBox.setStyles(
+                {
+                    height: '',
+                    width: ''
+                }
+            );
+        },
+
+        /**
+         * Set the dimensions of the video player based on the window size.
+         *
+         * @method _setResponsiveDimensions
+         * @protected
+         */
+        _setResponsiveDimensions: function() {
+            var aspectRatio,
+                currentTargetHeight,
+                currentTargetWidth,
+                instance,
+                height,
+                updatedHeight,
+                updatedWidth,
+                width,
+                winNode;
+
+            instance = this;
+
+            height = instance.get('height');
+            width = instance.get('width');
+
+            aspectRatio = height / width;
+
+            updatedHeight = height;
+            updatedWidth = width;
+
+            winNode = A.one(window);
+
+            currentTargetHeight = winNode.get('innerHeight');
+
+
+            if (currentTargetHeight < height) {
+                updatedHeight = currentTargetHeight;
+                updatedWidth = currentTargetHeight * aspectRatio;
+            }
+
+            currentTargetWidth = winNode.get('innerWidth');
+
+            if (currentTargetWidth < width) {
+                updatedHeight = currentTargetWidth * aspectRatio;
+                updatedWidth = currentTargetWidth;
+            }
+
+            instance._video.width(updatedWidth);
+            instance._video.height(updatedHeight);
         },
 
         /**